# backend/matchmaking_service/src/schemas/openapi.PublicAPI.yaml
openapi: 3.0.0
info:
  title: Public API
  version: 0.0.0
...
paths:
...
  /games/create:
    post:
      operationId: Games_create
      description: Online custom game request going to MM service. Body contains players that matchmaking service will invite via SSE.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameKey'
        '400':
          description: The server could not understand the request due to invalid syntax.
        '401':
          description: Access is unauthorized.
      tags:
        - Games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                numberOfPlayers:
                  type: integer
                  format: uint32
                  minimum: 2
                  maximum: 4
                  description: Capacity of the game. Host can invite as many as they like but starts when full.
                invitedPlayerIds:
                  type: array
                  items:
                    type: integer
                    format: uint64
              required:
                - numberOfPlayers
                - invitedPlayerIds
      security:
        - BearerAuth: []
...
  /games/{gameId}/decline:
    delete:
      operationId: Games_decline
      description: Decline game invitation.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '401':
          description: Access is unauthorized.
        '404':
          description: The server cannot find the requested resource.
      tags:
        - Games
      security:
        - BearerAuth: []
  /games/{gameId}/invite:
    post:
      operationId: Games_invitePlayers
      description: Invite players to a custom game - if not included in the initial create request.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteGameResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
        '401':
          description: Access is unauthorized.
        '404':
          description: The server cannot find the requested resource.
        '409':
          description: The request conflicts with the current state of the server.
      tags:
        - Games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invitedPlayerIds:
                  type: array
                  items:
                    type: integer
                    format: uint64
              required:
                - invitedPlayerIds
      security:
        - BearerAuth: []
  /games/{gameId}/join:
    post:
      operationId: Games_join
      description: Accept game invitation and receive token for WebSocket connection to game server.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameKey'
        '401':
          description: Access is unauthorized.
        '404':
          description: The server cannot find the requested resource.
        '409':
          description: The request conflicts with the current state of the server.
      tags:
        - Games
      security:
        - BearerAuth: []
  /games/{gameId}/view:
    post:
      operationId: Games_view
      description: Get viewing key for tournament game (spectator mode).
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  viewingKey:
                    type: string
                  gameId:
                    type: string
                required:
                  - viewingKey
                  - gameId
        '401':
          description: Access is unauthorized.
        '403':
          description: Access is forbidden.
        '404':
          description: The server cannot find the requested resource.
        '409':
          description: The request conflicts with the current state of the server.
      tags:
        - Games
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    GameKey:
      type: object
      required:
        - key
        - gameId
        - expires
      properties:
        key:
          type: string
        gameId:
          type: string
        expires:
          type: string
          format: date-time
      description: |-
        `GameKey` is returned by all the endpoints that are opening game websocket connection.
        Since the websocket endpoint cannot have body,
        `GameKey` will be the first message send via the newly established websocket"
...
    InviteGameResponse:
      type: object
      required:
        - gameId
        - invitedPlayers
      properties:
        gameId:
          type: string
        invitedPlayers:
          type: array
          items:
            type: integer
            format: uint64
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer